gossip: Call mark alive after on_join notification
In handle_major_state_change, we first mark a node alive then we call
the on_join callbacks which calls storage_service::on_change which in
turn calls storage_service::handle_state_replacing.

When the gossip status of the replacing node is STATUS_PREPARE_REPLACE,
which will be introduced in the following patch, we want to call
storage_service::handle_state_replacing before we mark a node as alive
to prevent CQL read code to include the replacing node to serve CQL
read. Because once an replacing node with the same IP address of the
dead node to be replaced is marked as alive by gossip, existing nodes
will start to include it as a target to serve reads.

Refs #7312