repair: Add strict mode for removenode ops
Currently, nodetool removenode operation runs in best effort mode which
means if a node that owns the range after the removenode operation is
down during the operation, the removenode node operation will continue
to succeed without requiring that node to perform data syncing. This can
cause data consistency issues. For example,

Five nodes in the cluster, RF = 3, for a range, n1, n2, n3 is the old
replicas, n2 is being removed, after the removenode operation, the new
replicas are n1, n5, n3. If n3 is down during the removenode operation,
only n1 will be used to sync data with the new owner n5. This will break
QUORUM read consistency if n1 happens to miss some writes.

To make removenode safe by default, we should fail the removenode
operation if the operation is not safe. If users want to continue the
unsafe removenode anyway, a best_effort_to_removenode option can be used
to switch to old best effort mode.

Fixes #7359