cdc: Add an "end-of-record" column to
Fixes #7435

Adds an "eor" (end-of-record) column to cdc log. This is non-null only on
last-in-timestamp group rows, i.e. end of a singular source "event".

A client can use this as a shortcut to knowing whether or not he has a
full cdc "record" for a given source mutation (single row change).

v2:
* Renamed eor column
* Removed optional state from log::builder, instead use batch_no as
  "state" (which it already is).
* Revert laxing of test check
v3:
* Fixed ck generation