utils: make updateable_value usable across shards
This patch makes `updateable_value` safe for cross-shard use for a far larger number of cases than before.
A proper solution would be to make the interface future-based and serialize the modifications to the vector of observers
via submit_to, but there exists a simpler idea:
1. The updateable value source, instead of having a single vector of observers, has a vector per shard.
2. Each observer tracks its owner shard.
3. When an observer is going to be removed, it knows which shard it came from and this information
   can be used to find it in the vector of observers per shard.

This mechanism prevents a most obvious race where multiple shards try to read and write to a single vector
residing on a single source (e.g. shard0). It's still possible to abuse the interface by reading from a value
from one shard and trying to remove it from another, but that would have to be explicitly coded,
which is rather unlikely to pass review. The second quite unlikely race would be updating the value at the same
time as adding/removing an observer, which requires a SIGHUP sent very carefully during startup or shutdown.