scylla_boot_setup: use kvm-clock for clock source on AWS Nitro VMs
We currently set 'clocksource=tsc' on bootparameter statically on our AMI,
since it is recommended for AWS Xen Hypervisor.
However, Amazon added AWS Nitro Hypervisor, which is recommended to use
kvm-clock as clocksource.
We need to support both hypervisor on single AMI image, so we need to configure
clocksource on first boot of the instance, not AMI creation type.

To do this, we need to detect which type of hypervisor we are on,
and set optimal clocksource to /sys/devices/system/clocksource/clocksource0/current_clocksource, not just change bootparameter settings.

Fixes #7444