utils: managed_bytes: introduce managed_bytes_view
managed_bytes_view is a non-owning view into managed_bytes,
but can also be converted from bytes_view and bytes.

Since it is a bytes_view replacement, it contains similar
methods (substr() and remove_prefix(), as used in the code
base) and a formatting operator. There is a maze of conversions
and comparison operators to avoid ambiguities when the compiler
needs to choose a path to arrive at the correct types.