types: abstract_type: add parallel interface with managed_bytes_view instead of bytes_view
For every method that takes a bytes_view, add a corresponding method
that takes a managed_bytes_view, so we can stop requiring that the
caller linearizes.

For now, the new implementations call the old ones, so we don't avoid any
linearizing, but we can later replace those implementations with
non-linearizing ones.

Once the transition is complete, we'll be able to drop the bytes_view
overloads.